:sparkles: jsonデータを生成できるように

+++ src/index.ts
@@ -3,21 +3,38 @@ import glob from "glob";
 import path from "path";
 import fs from "fs";
 import Settings from "../export_setting";
+import { JPCastles } from "./util";
+import { AreasFromPref, AreaType, Prefs, PrefType } from "./types/areas";
 
 const dirInit = (dirname: string) => {
-  fs.rmSync(dirname, { recursive: true });
+  if (fs.existsSync(dirname)) fs.rmSync(dirname, { recursive: true });
   fs.mkdirSync(dirname);
 };
 
-const addData = (searchDir: string, dataDir: string) => {
+const addDataByPref = (searchDir: string, dataDir: string) => {
   glob(path.join(searchDir, "/**/*"), { nodir: true }, (err, files) => {
     if (err) throw err;
 
     files.map((file) => {
       const newDir = path.dirname(file).replace(/^src\/data\//, `${dataDir}/`);
-      const newFile = path.basename(file, ".ts") + ".json";
+      const area = path.basename(file, ".ts");
+      const newFile = area + ".json";
       const newPath = `${newDir}/${newFile}`;
       const pathArr = newDir.split("/");
+
+      const pref = pathArr[2] as PrefType;
+      if (!Prefs.includes(pref))
+        throw new Error(
+          `フォルダ名が間違っています。  '${pref}' 'src/data/${pathArr[1]}/${pathArr[2]}'\n` +
+            " > npm run areas\n"
+        );
+      if (!(area in AreasFromPref))
+        throw new Error(
+          `フォルダ名が間違っています。 '${area}' 'src/data/${pathArr[1]}/${pathArr[2]}/${pathArr[3]}'\n` +
+            " > npm run areas"
+        );
+
+      const data = JPCastles[pref as PrefType][area as AreaType];
 
       pathArr.map((p, index) => {
         const currentPath = path.join(...pathArr.slice(0, index + 1));
@@ -26,13 +43,12 @@ const addData = (searchDir: string, dataDir: string) => {
         }
       });
 
-      fs.writeFile(newPath, newPath, (err) => {
+      fs.writeFile(newPath, JSON.stringify(data), (err) => {
         if (err) throw err;
-        console.log("正常に書き込みが完了しました");
       });
     });
   });
 };
 
-dirInit(Settings.export_dirname);
-addData("src/data", Settings.export_dirname);
+dirInit(Settings.export_dir);
+addDataByPref("src/data", Settings.export_dir);
